# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Rawaf Contractors Database
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Contractor {
  id: number;
  contractNo: string;
  contractor: string;
  project: string;
  portfolio: string;
  technicalScope: string;
  workType: string;
  price: number;
  phone: string;
  email: string;
  /** @nullable */
  workDescription?: string | null;
  /** @nullable */
  workScopeText?: string | null;
  /** @nullable */
  workCategory?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  businessProgram?: string | null;
  createdAt: string;
}

export interface CreateContractorBody {
  contractNo: string;
  contractor: string;
  project: string;
  portfolio: string;
  technicalScope: string;
  workType: string;
  price: number;
  phone: string;
  email: string;
  /** @nullable */
  workDescription?: string | null;
  /** @nullable */
  workScopeText?: string | null;
  /** @nullable */
  workCategory?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  businessProgram?: string | null;
}

export interface Contract {
  id: number;
  contractNo: string;
  title: string;
  vendorName: string;
  vendorContact: string;
  value: number;
  startDate: string;
  endDate: string;
  contractType: string;
  projectName: string;
  currentStage: number;
  status: string;
  createdBy: string;
  /** @nullable */
  wordFilename?: string | null;
  /** @nullable */
  signedFilename?: string | null;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateContractBody {
  title: string;
  vendorName: string;
  vendorContact?: string;
  value?: number;
  startDate?: string;
  endDate?: string;
  contractType?: string;
  projectName?: string;
  createdBy?: string;
}

export interface ContractStats {
  total: number;
  draft: number;
  inProgress: number;
  approved: number;
  completed: number;
}

export interface ContractStageLog {
  id: number;
  contractId: number;
  stage: number;
  action: string;
  actorRole: string;
  actorName: string;
  notes: string;
  createdAt: string;
}

export interface ContractDocument {
  id: number;
  contractId: number;
  stage: number;
  filename: string;
  fileType: string;
  uploadedBy: string;
  createdAt: string;
}

export interface AddDocumentBody {
  stage: number;
  filename: string;
  fileType?: string;
  uploadedBy?: string;
}

export type StageActionBodyAction =
  (typeof StageActionBodyAction)[keyof typeof StageActionBodyAction];

export const StageActionBodyAction = {
  advance: "advance",
  reject: "reject",
} as const;

export interface StageActionBody {
  action: StageActionBodyAction;
  actorRole: string;
  actorName: string;
  notes?: string;
  wordFilename?: string;
  signedFilename?: string;
}

export interface ContractPdfData {
  contract: Contract;
  auditLog: ContractStageLog[];
  documents: ContractDocument[];
}

export type ListContractorsParams = {
  contractNo?: string;
  contractor?: string;
  technicalScope?: string;
  workType?: string;
  project?: string;
  portfolio?: string;
};

export type ListContractsParams = {
  status?: string;
  stage?: number;
};