# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Rawaf Contractors Database
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contractors
    description: Contractor and supplier records
  - name: contracts
    description: Contract management lifecycle (11-stage workflow)

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /contractors:
    get:
      operationId: listContractors
      tags: [contractors]
      summary: List all contractors
      parameters:
        - name: contractNo
          in: query
          schema: { type: string }
        - name: contractor
          in: query
          schema: { type: string }
        - name: technicalScope
          in: query
          schema: { type: string }
        - name: workType
          in: query
          schema: { type: string }
        - name: project
          in: query
          schema: { type: string }
        - name: portfolio
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of contractors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contractor"

    post:
      operationId: createContractor
      tags: [contractors]
      summary: Create a new contractor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractorBody"
      responses:
        "201":
          description: Created contractor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contractor"
        "400":
          description: Validation error

  /contractors/{id}:
    get:
      operationId: getContractor
      tags: [contractors]
      summary: Get a contractor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Contractor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contractor"
        "404":
          description: Not found

    put:
      operationId: updateContractor
      tags: [contractors]
      summary: Update a contractor
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractorBody"
      responses:
        "200":
          description: Updated contractor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contractor"
        "400":
          description: Validation error
        "404":
          description: Not found

    delete:
      operationId: deleteContractor
      tags: [contractors]
      summary: Delete a contractor
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Not found

  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List all contracts
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: stage
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contract"

    post:
      operationId: createContract
      tags: [contracts]
      summary: Create a new contract (stage 1)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContractBody"
      responses:
        "201":
          description: Created contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "400":
          description: Validation error

  /contracts/stats:
    get:
      operationId: getContractStats
      tags: [contracts]
      summary: Get contract KPI statistics
      responses:
        "200":
          description: Contract statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractStats"

  /contracts/{id}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get a single contract by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Contract detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found

  /contracts/{id}/stage:
    patch:
      operationId: advanceContractStage
      tags: [contracts]
      summary: Advance or reject a contract stage (role-gated)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageActionBody"
      responses:
        "200":
          description: Updated contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "400":
          description: Validation error
        "403":
          description: Role not authorized for this stage
        "404":
          description: Not found
        "409":
          description: Contract already completed

  /contracts/{id}/audit:
    get:
      operationId: getContractAudit
      tags: [contracts]
      summary: Get full audit trail for a contract
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContractStageLog"

  /contracts/{id}/documents:
    get:
      operationId: getContractDocuments
      tags: [contracts]
      summary: List documents attached to a contract
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContractDocument"

    post:
      operationId: addContractDocument
      tags: [contracts]
      summary: Attach a document record to a contract stage
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDocumentBody"
      responses:
        "201":
          description: Document record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractDocument"
        "400":
          description: Validation error

  /contracts/{id}/pdf-data:
    get:
      operationId: getContractPdfData
      tags: [contracts]
      summary: Get full contract data for PDF generation (contract + audit + documents)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Full PDF data bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContractPdfData"
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Contractor:
      type: object
      properties:
        id:
          type: integer
        contractNo:
          type: string
        contractor:
          type: string
        project:
          type: string
        portfolio:
          type: string
        technicalScope:
          type: string
        workType:
          type: string
        price:
          type: integer
        phone:
          type: string
        email:
          type: string
        workDescription:
          type: ["string", "null"]
        workScopeText:
          type: ["string", "null"]
        workCategory:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        businessProgram:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - contractNo
        - contractor
        - project
        - portfolio
        - technicalScope
        - workType
        - price
        - phone
        - email
        - createdAt

    CreateContractorBody:
      type: object
      properties:
        contractNo:
          type: string
        contractor:
          type: string
        project:
          type: string
        portfolio:
          type: string
        technicalScope:
          type: string
        workType:
          type: string
        price:
          type: integer
        phone:
          type: string
        email:
          type: string
        workDescription:
          type: ["string", "null"]
        workScopeText:
          type: ["string", "null"]
        workCategory:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        businessProgram:
          type: ["string", "null"]
      required:
        - contractNo
        - contractor
        - project
        - portfolio
        - technicalScope
        - workType
        - price
        - phone
        - email

    Contract:
      type: object
      properties:
        id:
          type: integer
        contractNo:
          type: string
        title:
          type: string
        vendorName:
          type: string
        vendorContact:
          type: string
        value:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        contractType:
          type: string
        projectName:
          type: string
        currentStage:
          type: integer
        status:
          type: string
        createdBy:
          type: string
        wordFilename:
          type: ["string", "null"]
        signedFilename:
          type: ["string", "null"]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - contractNo
        - title
        - vendorName
        - vendorContact
        - value
        - startDate
        - endDate
        - contractType
        - projectName
        - currentStage
        - status
        - createdBy
        - createdAt
        - updatedAt

    CreateContractBody:
      type: object
      properties:
        title:
          type: string
        vendorName:
          type: string
        vendorContact:
          type: string
        value:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        contractType:
          type: string
        projectName:
          type: string
        createdBy:
          type: string
      required:
        - title
        - vendorName

    ContractStats:
      type: object
      properties:
        total:
          type: integer
        draft:
          type: integer
        inProgress:
          type: integer
        approved:
          type: integer
        completed:
          type: integer
      required:
        - total
        - draft
        - inProgress
        - approved
        - completed

    ContractStageLog:
      type: object
      properties:
        id:
          type: integer
        contractId:
          type: integer
        stage:
          type: integer
        action:
          type: string
        actorRole:
          type: string
        actorName:
          type: string
        notes:
          type: string
        createdAt:
          type: string
      required:
        - id
        - contractId
        - stage
        - action
        - actorRole
        - actorName
        - notes
        - createdAt

    ContractDocument:
      type: object
      properties:
        id:
          type: integer
        contractId:
          type: integer
        stage:
          type: integer
        filename:
          type: string
        fileType:
          type: string
        uploadedBy:
          type: string
        createdAt:
          type: string
      required:
        - id
        - contractId
        - stage
        - filename
        - fileType
        - uploadedBy
        - createdAt

    AddDocumentBody:
      type: object
      properties:
        stage:
          type: integer
        filename:
          type: string
        fileType:
          type: string
        uploadedBy:
          type: string
      required:
        - stage
        - filename

    StageActionBody:
      type: object
      properties:
        action:
          type: string
          enum: [advance, reject]
        actorRole:
          type: string
        actorName:
          type: string
        notes:
          type: string
        wordFilename:
          type: string
        signedFilename:
          type: string
      required:
        - action
        - actorRole
        - actorName

    ContractPdfData:
      type: object
      properties:
        contract:
          $ref: "#/components/schemas/Contract"
        auditLog:
          type: array
          items:
            $ref: "#/components/schemas/ContractStageLog"
        documents:
          type: array
          items:
            $ref: "#/components/schemas/ContractDocument"
      required:
        - contract
        - auditLog
        - documents