# lib/api-zod/src/generated/types/addDocumentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Rawaf Contractors Database
 * OpenAPI spec version: 0.1.0
 */

export interface AddDocumentBody {
  stage: number;
  filename: string;
  fileType?: string;
  uploadedBy?: string;
}