# lib/api-zod/src/generated/types/contract.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Rawaf Contractors Database
 * OpenAPI spec version: 0.1.0
 */

export interface Contract {
  id: number;
  contractNo: string;
  title: string;
  vendorName: string;
  vendorContact: string;
  value: number;
  startDate: string;
  endDate: string;
  contractType: string;
  projectName: string;
  currentStage: number;
  status: string;
  createdBy: string;
  /** @nullable */
  wordFilename?: string | null;
  /** @nullable */
  signedFilename?: string | null;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: string;
  updatedAt: string;
}