# lib/api-zod/src/generated/types/contractDocument.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Rawaf Contractors Database
 * OpenAPI spec version: 0.1.0
 */

export interface ContractDocument {
  id: number;
  contractId: number;
  stage: number;
  filename: string;
  fileType: string;
  uploadedBy: string;
  createdAt: string;
}