# lib/api-zod/src/generated/types/contractPdfData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Rawaf Contractors Database
 * OpenAPI spec version: 0.1.0
 */
import type { Contract } from "./contract";
import type { ContractDocument } from "./contractDocument";
import type { ContractStageLog } from "./contractStageLog";

export interface ContractPdfData {
  contract: Contract;
  auditLog: ContractStageLog[];
  documents: ContractDocument[];
}