# lib/api-zod/src/generated/types/contractStageLog.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Rawaf Contractors Database
 * OpenAPI spec version: 0.1.0
 */

export interface ContractStageLog {
  id: number;
  contractId: number;
  stage: number;
  action: string;
  actorRole: string;
  actorName: string;
  notes: string;
  createdAt: string;
}