# lib/api-zod/src/generated/types/contractStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Rawaf Contractors Database
 * OpenAPI spec version: 0.1.0
 */

export interface ContractStats {
  total: number;
  draft: number;
  inProgress: number;
  approved: number;
  completed: number;
}