# lib/api-zod/src/generated/types/contractor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Rawaf Contractors Database
 * OpenAPI spec version: 0.1.0
 */

export interface Contractor {
  id: number;
  contractNo: string;
  contractor: string;
  project: string;
  portfolio: string;
  technicalScope: string;
  workType: string;
  price: number;
  phone: string;
  email: string;
  /** @nullable */
  workDescription?: string | null;
  /** @nullable */
  workScopeText?: string | null;
  /** @nullable */
  workCategory?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  businessProgram?: string | null;
  createdAt: string;
}