# lib/api-zod/src/generated/types/createContractBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Rawaf Contractors Database
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContractBody {
  title: string;
  vendorName: string;
  vendorContact?: string;
  value?: number;
  startDate?: string;
  endDate?: string;
  contractType?: string;
  projectName?: string;
  createdBy?: string;
}