# lib/api-zod/src/generated/types/stageActionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Rawaf Contractors Database
 * OpenAPI spec version: 0.1.0
 */
import type { StageActionBodyAction } from "./stageActionBodyAction";

export interface StageActionBody {
  action: StageActionBodyAction;
  actorRole: string;
  actorName: string;
  notes?: string;
  wordFilename?: string;
  signedFilename?: string;
}